refactor(defaults): use lodash defaults instead of custom utilities
Lodash is a highly tested, and often used, library containing a number of
very useful methods some of which have been duplicated in this library's
utility class. This change introduced lodash as a dependency, and opts
for using its methods over the custom implementations. For starters, each
frame has been converted to using lodash.defaults.