refactor(disposition): reduced duplication, added error handling
There was a certain amount of duplication in the disposition frame
handling which was easily consolidated. Also, testing against a QPID
server exposed some issues with rejected messages. I've left a comment
in there for future discussion about what should be returned in the
event that a rejection error is not provided. Note that a few debug
messages were temporarily added to help trace errors in the meantime
before a general debug cleanup occurs.