refactor(frame_reader): reduce complixity and enhance readability
FrameReader used to be a large collection of switch statements. This
change reduces that complexity by splitting into the concept of AMQP
and SASL frame internal reader classes. A number of other readability
cleanups occurred including: unrolling a number of if statements to
fail early, as well as registration of Readers.