refactor(test-send-buffers): streamline outgoing buffers for tests
A few things going on here:

 - remove all the temporary <frameType>buf() methods from tests,
   leading to a more declarative form for what the expected
   requests and responses are for a test sequence
 - removal of the Frame.outgoing helper function

This is the first step towards removing the buffer-builder from
the codebase. Since we are working with streams, there should be
no need to actually use a buffer builder. Now that this is in
place, we are able to use modules like buffer-more-ints to work
directly with streams (initially) for outgoing encoding.