refactor(timestamps): use Date as the primary expression of ms64
Timestamps in AMQP 1.0 are expressed as 64bit unsigned integers, of
course this doesn't jive with Javascript. As a convenience to our
users, this patch defaults to converting all timestamps (incoming
and outgoing) to Date objects. This can be imprecise, and we will
need to investigate future means of opting out of this conversion.