fix(SenderLink): temporarily remove transfersInFlight
Session._transfersInFlight was used to keep track of messages that
may have been required to be resent upon link reattachment. As we
currently have no logic for this (and this was being set whether we
were guaranteed to get disposition for the frame or not), I've
opted to remove this until that behavior is more completely
implemented.