feat(Client): remove receipt callback on createReceiver
This greatly simplifies the internals for receiver link creation,
aligning it more closely with sender link creation (so we can begin
to share code between them). It also makes the API more explicit by
requiring the user to understand that the receiver link is indeed
an EventEmitter and should be used that way.