feat(unique-link-names): use uniquely generated link names
This follows more closely to what QPID is doing with their
messaging framework, in that link names by default are generated
based on their address (and in our case link creation properties).
A hash of the address and options is used to uniquely identify an
attempt to recreate an existing receiver. The user is still able
to override this behavior by passing in a "name" parameter to the
link creation "options" object (verified in the unit tests)