fix(link-creation): remove support for tracking duplicate links
Per our discussion in the gitter channel, support for tracking
attempts to create duplicate links has been fully removed. This
greatly reduces the complexity of link creation, allowing us to
remove Session._linksByName, the need for an object hash, as well
as a swath of logic to track links by name/id/options.