refactor(timestamps): enable fully supported range of numbers
JavaScript supports up to 2^53 in its Number type, however for
timestamps we were cutting off support at 2^32 - 1. This is because
we can potentially lose precision with sufficiently large numbers,
however it's a bit restrictive to reduce the supported numbers to
just 2^32 - 1. We now will silently drop precision given large
enough numbers, which may be less desirable (and can be fixed!),
but the user experience is much better now as they can use e.g.
Date.now() for something like "absoluteExpiryTime"

closes #177