fix(client): connect promise should resolve if reconnect is needed
Our reconnect logic was only ever tested once a connection already
existed. Unfortunately, that meant that we missed the reconnection
case when a reconnect was required from the very start of the
attempt. The connect promise was not being resolved in this case,
so that has now been remedied and two unit tests have been added
to ensure the behavior persists.

closes #190