test(streams): move SB streams test to EH to avoid conflicts
When we test against just a service bus queue, the test are run in
a race since they are all using the same bus for testing. This has
lead to flakey tests, and all out CI test failure. Instead we test
the feature using EventHubs which allow for more deterministic
behavior.