fix(defaultSubjects): throw error on invalid subjects
Its possible (and perhaps very likely) for default subjects to be
specified using a variable which might be undefined, and therefore
coerced to a string value as such. As a result, the user will
likely be using an unexpected link unknowingly. Instead, we will
now throw an error on these invalid cases.