refactor(CompositeType): use setters for incoming described types
A missing piece of the type system refactor was that when described
types came in, we simply set the underlaying value of the composite
type to the value of the described type. This results in loss of
type information, specifically related to defaults and other
constraints.

This patch fixes that problem, but potentially introduces others.
You might notice that some less-than-ideal changes were introduced
around symbol's, and having to explicitly use `ForcedType`. This is
because we simply can't overload the comparison operators allowing
`ForcedType`s to equal their (e.g.) `String` counterparts.