feat(connection): terminate connection when frame steam interrupted
Previously, if invalid data was encountered while processing
incoming frames we would silently accept the data and continue on
our merry way - except we wouldn't process any future data. This
patch introduces a rather harsh assumption that incoming junk data
indicates a bad connection, and therefore terminates the connection
It could be argued that in the future we simply throw the junk data
away and wait for more frames, though it remains to be seen who
would actually want that functionality.